# Conflicts:
#	unittest/force-styles/tests/mol-pair-born.yaml
#	unittest/force-styles/tests/mol-pair-born_coul_dsf.yaml
#	unittest/force-styles/tests/mol-pair-born_coul_long.yaml
#	unittest/force-styles/tests/mol-pair-born_coul_msm.yaml
#	unittest/force-styles/tests/mol-pair-born_coul_wolf.yaml
#	unittest/force-styles/tests/mol-pair-buck.yaml
#	unittest/force-styles/tests/mol-pair-buck_coul_long.yaml
#	unittest/force-styles/tests/mol-pair-buck_coul_msm.yaml
#	unittest/force-styles/tests/mol-pair-lj_cubic.yaml
#	unittest/force-styles/tests/mol-pair-morse.yaml
#	unittest/force-styles/tests/mol-pair-tip4p_cut.yaml
#	unittest/force-styles/tests/mol-pair-zbl.yaml

+++ src/USER-INTEL/nbin_intel.h
@@ -40,7 +40,7 @@ class NBinIntel : public NBinStandard {
   FixIntel *_fix;
   int *_atombin, *_binpacked;
   int _precision_mode;
-  bigint memory_usage();
+  double memory_usage();
 
   template <class flt_t, class acc_t>
   void bin_atoms(IntelBuffers<flt_t,acc_t> *);
